add params for antenna pose
getting the transformation from the tf transformation did not work, we where getting error that it could not find the transformation between the two links! however when we were trying to get the transformation from the terminal using this command: "ros2 run tf2_ros tf2_echo back_antenna_link front_antenna_link" we were able to get it! so as a fix for "now" we hardcoded those params!!!!!